Fixed the dot product function to calculate matrices.
I did a lot of testing and debugging in order to solve the error, which is simple and obvious when you think about it. I also made a python program to compare the results with the numpy library.

+++ kinematics/spot_micro_kinematics.cc
@@ -11,8 +11,8 @@ template<typename T, typename A> struct is_std_vector<vector<T, A>> : true_type 
 template <typename T>
 enable_if_t<is_std_vector<decay_t<T>>::value, T>
 operator+(T&& vec1, T&& vec2) {
-	for (int i = 0; i < vec2.size(); ++i) {
-    for (int j = 0; j < vec2[i].size(); j++) {
+	for (size_t i = 0; i < vec2.size(); ++i) {
+    for (size_t j = 0; j < vec2[i].size(); j++) {
       vec1[i][j] += vec2[i][j];
     }
   }
